If else and sesibility thrash hold

See merge request ti6120/MetroFlutterVSC!7

+++ metrolyse/metrolyse/lib/model/metronome_funktion.dart
@@ -20,6 +20,7 @@ class MetronomeFunctionState extends State<MetronomeFunction> {
   //Timer class
   ReliableIntervalTimer? _clickTimer;
 
+//Audio Functions from the Class AudioPlay
   AudioPlay audioPlay = AudioPlay();
 
 //ContextBuilder
@@ -62,7 +63,6 @@ class MetronomeFunctionState extends State<MetronomeFunction> {
         interval: Duration(milliseconds: min ~/ bpmInit),
         callback: (int elapsedMilliseconds) async {
           await audioPlay.playClick();
-          // SystemSound.play(SystemSoundType.click);
           getMetroDates();
           checkAlgo.metroDates();
         });
@@ -70,24 +70,22 @@ class MetronomeFunctionState extends State<MetronomeFunction> {
 
   // Starts the Metronome-Timer
   startClick() async {
-    // _clickTimer?.stop();
     _clickTimer?.start();
-    // stopClick();
   }
 
 // Sets the click or acceleration Date
   getMetroDates() {
-    double clickDate = DateTime.now().millisecondsSinceEpoch.toDouble();
+    int clickDate = DateTime.now().millisecondsSinceEpoch;
     return clickDate;
   }
 
+// Updates the Metronome
   updateClick() {
     _clickTimer?.updateInterval(Duration(milliseconds: min ~/ bpmInit));
   }
 
-  // //Stops the Metronome-Timer
+  // Stops the Metronome-Timer
   void stopClick() async {
-    // audioPlay.muteClick();
     await _clickTimer?.stop();
   }
